Remove the ability to use pytorch device `cpu`.
Using a CPU for ESRGAN is generally unsafe.

- It takes incredibly long and most likely a waste of time.
- Will cause extremely high temps and max out CPU utilization, risking the CPU to damage over time.

In general, it's like smoking, people do it but they shouldnt. To help stop people (as most probably don't know they shouldn't be doing this) I have forced the inability to use it.

If someone knows what they are doing, and they really do want to use CPU, then they can use `cpu!` as the device argument.