fix: Remove plane access check
Closes #9

Thanks to @JoeyBallentine for the majority of the code that went into the support for this.

It only splits into quadrants one time (so 4 total quadrants) for now. In the future I may add support for further quadrants but for now this will definitely do.

It's seamless as it extends the quadrants edges (except for outer image edges) a small amount so that when the model gets ran, it will have the expected artefacts on the edges, which will be then cropped off. When then merged back together it is literally seamless.

The overlap integer is what controls this behavior. Setting overlap to 0 or a false-y value will disable chunking altogether. Settings it to a value > 1 will extend each quadrant by that amount. It's generally recommended to have this extension be large.

In my testing on a 720x576 input, I saw seams with it at 8 or lower, though at 8 it was incredibly hard to notice. It's generally recommended to go no less than 16, and you should raise that the larger your input resolution * model scale is.

So, for 720x576 with a 2x model scale, 16 worked fine for my source and my model. Yours may differ. If you want to be safe, just go 16 and if you notice even a bit of a seam, double it until you dont.