Add self-chaining support
This makes it so functions return it's own self instance so that you can immediately call another one of that instance's objects of properties.

This makes it super easy and semantic to chain calls for using multiple models, or running the same one multiple times.

For example:
```py
clip = VSGAN(0).\
    load_model(r"C:\Users\PHOENiX\Downloads\2X_DigitalFilmV5_Lite.pth").\
    run(clip, overlap=0).\
    load_model(r"C:\Users\PHOENiX\Downloads\4X_Test.pth").\
    run(clip, overlap=16).\
    load_model(r"C:\Users\PHOENiX\Downloads\4X_DoubleRunExample.pth").\
    run(clip, overlap=16).\
    run(clip, overlap=16).\
    clip
```