docs: Remove Stable/Latest links from sidebar
These links don't realistically work on readthedocs.io mainly due to the /{lang}/ prefix on the urls. Even if I were to fix it by detecting the lang, whats the point when there's a button on the bottom right of readthedocs page's that let you choose stable or latest as well as a LOT more too.