Rework model file sys structure, Rework ESRGAN Module as one class
This is essentially just a few moved around files, with the sanitize_state_dict (which was always ESRGAN specific) to only be used by ESRGAN, but without having to use any `if` checks or anything.

Its now used directly within ESRGANs network class, which is renamed from RRDBNet to ESRGAN. It's also been heavily cleaned up to be all in the one call but readable.

This also heavily simplifies the call to the model creation in VSGANs main code, and opens up the possibility for easier importing of other Architecture's Networks.