Optimize frame_to_tensor args, docstring, and code
This removes add_batch as I felt it was unnecessary due to its simplicity, and the fact this is mostly applied *after* the conversion to a tensor has occurred. The other post-operations are kept due to their complexity or convenience-sake.

Use of np.ascontiguousarray didnt seem necessary, so it was removed.

The parameter change_order was replaced with order, which can specify the exact output order you want. If you leave it at the default, it wont change order.