Separate VSGAN class into ESRGAN & EGVSR classes
Rework `load_model` to just `load`, and `run` to `apply`.

This change is to simplify the import paths for current and future networks. Since VSGAN is starting to become more than just ESRGAN networks, its becoming complicated code wise and UX-wise on how to execute a model, and what code would run.

With this change we go from using a vague `VSGAN` class, to separate classes for each network.