Simplify VideoFrame memory access in frame_to_tensor
Originally I expected that I'd need to explicitly convert the possibility of unsigned formats to signed data, but that does not seem to be the case.

All formats are returning in their signed format, possibly because of dtype's frombuffer specifying signed formats, it may be casting to the signed format.