Add Multi-GPU processing in an alternative way
This is similar to `multi-gpu` branch but at a higher level using more VapourSynth than Torch code.

This method isn't ideal but effectively boils down to splitting the video into 2 clips, first is every n+0 frames, second is every n+1 frames.

It then runs the model on each with different devices, and it then merges the two clips back together.

For whatever reason this actually works to get a performance increase compared to `multi-gpu` which effectively refuses to work on frames in parallel.