Add support for SwinIR
Note: SwinIR arch code does not currently support float16 (HalfTensor), so any input clip that is float16 will be cast to float32.

Weight Initialization funcs are used in SwinIR which imports the needed functions from torch. This has required a lift of the minimum supported PyTorch version from v1.10.0 to v1.12.0. This is still a relatively old version so support with other environments and tools shouldn't be a problem.

Closes #10