docs: Don't indent Real-ESRGAN as part of ESRGAN
While Real-ESRGAN v1 is ESRGAN with a on-the-fly HR->LR degradation model, Real-ESRGAN v2 (as most people know and use it nowadays) is a completely different architecture called SRVGGNetCompact, usually referred to as "Compact".

What's strange about Real-ESRGAN (v2) is that the latest arxiv paper about Real-ESRGAN still implies the use of RRDBNet (ESRGAN), yet the official repositories for Real-ESRGAN has moved to SRVGGNetCompact.